careerflow-agent-core/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-docx
pdfplumber
pydantic
sqlalchemy
python-dotenv
langchain
langchain-groq
langchain-community
chromadb
sentence-transformers
firebase-admin
